Adding module randomdict.
This is a fork of the original available at https://github.com/emeryberger/randomdict/blob/master/randomdict.py.
It fixes the import collections that needs to be import collections.abc in the original randomdict module. Without that fix, we can't pip install the module with recent versions of python 3.10+
So, grabbed the whole thing and will import it locally instead.